app/requirements.txt
flask
flask-cors
flask_sqlalchemy
flask_login
gunicorn
requests
Werkzeug
markdown
PyYAML
imageio
imageio-ffmpeg
matplotlib
bs4
python-dotenv